LayoutTests/fast/page-color-sampling/color-sampling-in-position-sticky.html

<!DOCTYPE html> <!-- webkit-test-runner [ ContentInsetBackgroundFillEnabled=true pageTopColorSamplingEnabled=true useFlexibleViewport=true obscuredInset.top=100 ] -->
<html>
<head>
    <meta charset="utf-8">
    <meta name="viewport" content="width=device-width, initial-scale=1">
    <style>
    body, html {
        width: 100%;
        height: 100%;
        margin: 0;
        font-family: system-ui;
        box-sizing: border-box;
        border: 0.5px solid lightgray;
    }

    footer {
        position: absolute;
        bottom: 0;
        width: 100%;
        height: 60px;
        background: darkgray;
        color: white;
        text-align: center;
    }

    header {
        position: sticky;
        top: 0;
        width: 100%;
        height: 60px;
        z-index: 100;
    }

    .header-content {
        position: absolute;
        top: 0;
        width: 100%;
        height: 100%;
        background: rgb(250, 60, 0);
        will-change: transform;
    }

    .tall {
        width: 10px;
        height: 5000px;
    }
    </style>
    <script src="../../resources/js-test.js"></script>
    <script src="../../resources/ui-helper.js"></script>
    <script>
    jsTestIsAsync = true;

    addEventListener("load", async () => {
        await UIHelper.ensurePresentationUpdate();
        colorsBeforeScrolling = await UIHelper.fixedContainerEdgeColors();
        shouldBeEqualToString("colorsBeforeScrolling.top", "rgb(250, 60, 0)");
        shouldBeNull("colorsBeforeScrolling.left");
        shouldBeNull("colorsBeforeScrolling.right");
        shouldBeNull("colorsBeforeScrolling.bottom");

        scrollBy(0, 5000);
        await UIHelper.ensurePresentationUpdate();

        colorsAfterScrolling = await UIHelper.fixedContainerEdgeColors();
        shouldBeEqualToString("colorsAfterScrolling.top", "rgb(250, 60, 0)");
        shouldBeNull("colorsAfterScrolling.left");
        shouldBeNull("colorsAfterScrolling.right");
        shouldBeNull("colorsAfterScrolling.bottom");
        finishJSTest();
    });
    </script>
</head>
<body>
    <header>
        <div class="header-content"></div>
    </header>
    <footer>No sticky header after this point</footer>
    <div class="tall"></div>
</body>
</html>
