LayoutTests/fast/page-color-sampling/color-sampling-in-position-sticky-expected.txt

PASS colorsBeforeScrolling.top is "rgb(250, 60, 0)"
PASS colorsBeforeScrolling.left is null
PASS colorsBeforeScrolling.right is null
PASS colorsBeforeScrolling.bottom is null
PASS colorsAfterScrolling.top is "rgb(250, 60, 0)"
PASS colorsAfterScrolling.left is null
PASS colorsAfterScrolling.right is null
PASS colorsAfterScrolling.bottom is null
PASS successfullyParsed is true

TEST COMPLETE
No sticky header after this point
